Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib
aiomysql
alembic